Added saving and removing games to and from favorites Closes #9
