+++ GMAPP/Models/Game.swift
@@ -25,26 +25,6 @@ struct Game : Codable, Identifiable {
     static let example = Game(id: 12312, title: "Mass Effect Andromeda", platform: "PC", release_date: "2021-05-14T00:00:00.000Z")
     #endif
     
-    static func encodeGamesToDefaults(games: [Game]) {
-        let encoder = JSONEncoder()
-        if let encoded = try? encoder.encode(games) {
-            let defaults = UserDefaults.standard
-            defaults.set(encoded, forKey: "favGames")
-        }
-    }
-    
-    static func decodeGamesToDefaults() -> [Game] {
-        var favouriteGames = [Game]()
-        let defaults = UserDefaults.standard
-        if let favGames = try? defaults.object(forKey: "favGames") as? Data {
-            let decoder = JSONDecoder()
-            if let loadedGames = try? decoder.decode([Game].self, from: favGames) {
-                Logger.shared.log("Loaded games for UserDefaults: \(loadedGames.count)")
-                favouriteGames = loadedGames
-            }
-        }
-        return favouriteGames ?? [Game.example]
-    }
     
 }
